requirements.txt
aerich==0.7.2 ; python_version >= "3.11" and python_version < "4.0"
aiosqlite==0.17.0 ; python_version >= "3.11" and python_version < "4.0"
annotated-types==0.6.0 ; python_version >= "3.11" and python_version < "4.0"
anyio==3.7.1 ; python_version >= "3.11" and python_version < "4.0"
bcrypt==4.0.1 ; python_version >= "3.11" and python_version < "4.0"
bidict==0.22.1 ; python_version >= "3.11" and python_version < "4.0"
click==8.1.7 ; python_version >= "3.11" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.11" and python_version < "4.0" and (sys_platform == "win32" or platform_system == "Windows")
dictdiffer==0.9.0 ; python_version >= "3.11" and python_version < "4.0"
fastapi==0.104.1 ; python_version >= "3.11" and python_version < "4.0"
h11==0.14.0 ; python_version >= "3.11" and python_version < "4.0"
httptools==0.6.1 ; python_version >= "3.11" and python_version < "4.0"
idna==3.4 ; python_version >= "3.11" and python_version < "4.0"
iso8601==1.1.0 ; python_version >= "3.11" and python_version < "4.0"
markdown-it-py==3.0.0 ; python_version >= "3.11" and python_version < "4.0"
mdurl==0.1.2 ; python_version >= "3.11" and python_version < "4.0"
pydantic-core==2.14.3 ; python_version >= "3.11" and python_version < "4.0"
pydantic==2.5.1 ; python_version >= "3.11" and python_version < "4.0"
pygments==2.16.1 ; python_version >= "3.11" and python_version < "4.0"
pypika-tortoise==0.1.6 ; python_version >= "3.11" and python_version < "4.0"
python-dotenv==1.0.0 ; python_version >= "3.11" and python_version < "4.0"
python-engineio==4.8.1 ; python_version >= "3.11" and python_version < "4.0"
python-socketio==5.10.0 ; python_version >= "3.11" and python_version < "4.0"
pytz==2023.3.post1 ; python_version >= "3.11" and python_version < "4.0"
pyyaml==6.0.1 ; python_version >= "3.11" and python_version < "4.0"
readchar==4.0.5 ; python_version >= "3.11" and python_version < "4.0"
rich-click==1.7.2 ; python_version >= "3.11" and python_version < "4.0"
rich==13.7.0 ; python_version >= "3.11" and python_version < "4.0"
setuptools==68.2.2 ; python_version >= "3.11" and python_version < "4.0"
simple-websocket==1.0.0 ; python_version >= "3.11" and python_version < "4.0"
sniffio==1.3.0 ; python_version >= "3.11" and python_version < "4.0"
starlette==0.27.0 ; python_version >= "3.11" and python_version < "4.0"
tomlkit==0.12.3 ; python_version >= "3.11" and python_version < "4.0"
tortoise-orm==0.20.0 ; python_version >= "3.11" and python_version < "4.0"
typing-extensions==4.8.0 ; python_version >= "3.11" and python_version < "4.0"
uvicorn[standard]==0.24.0.post1 ; python_version >= "3.11" and python_version < "4.0"
uvloop==0.19.0 ; (sys_platform != "win32" and sys_platform != "cygwin") and platform_python_implementation != "PyPy" and python_version >= "3.11" and python_version < "4.0"
watchfiles==0.21.0 ; python_version >= "3.11" and python_version < "4.0"
websockets==12.0 ; python_version >= "3.11" and python_version < "4.0"
wsproto==1.2.0 ; python_version >= "3.11" and python_version < "4.0"
ascenderai==0.1.0 ; python_version >= "3.11" and python_version < "4.0"